+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## dev
+
+- Add example implementing a pausable countdown.
+
 ## 0.1.0+1
 
 - Add build status and sponsorship badges.